extract stuff from main file
