+++ main.go
@@ -3,12 +3,10 @@ package main
 import (
 	"DocPlanner/pingdom-statuspage-integration/statuspage"
 	"fmt"
-	"net/http"
 	"os"
 	"strconv"
 	"time"
 )
-import "github.com/gin-gonic/gin"
 
 type Response struct {
 	Status string `json:"status"`
@@ -19,18 +17,7 @@ func main() {
 	defer ticker.Stop()
 
 	componentsStoreChan := make(chan *componentsStore)
-	go func() {
-		var cs *componentsStore
-		for {
-			select {
-			case a := <-componentsStoreChan:
-				cs = a
-			case <-ticker.C:
-				err := cs.Refresh()
-				fmt.Println(fmt.Sprintf("[%s] Refreshing StatusPage components state! %s", time.Now().Format(time.RFC1123Z), err.Error()))
-			}
-		}
-	}()
+	go AsyncRefresh(ticker, componentsStoreChan)
 
 	secret := getSecret()
 	statusPageClient := setupStatusPageClient()
@@ -72,51 +59,4 @@ func setupStatusPageClient() *statuspage.Client {
 	}
 
 	return statusPageClient
-}
-
-func InitializeComponentsStore(statusPageClient *statuspage.Client, componentsStoreChan chan *componentsStore) gin.HandlerFunc {
-	componentStore := NewComponentsStore(statusPageClient)
-	err := componentStore.Refresh()
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(2)
-	}
-
-	if componentsStoreChan != nil {
-		componentsStoreChan <- componentStore
-	}
-
-	return func(context *gin.Context) {
-		context.Set("component_store", componentStore)
-		context.Next()
-	}
-}
-
-func BananaAuthMiddleware(secret string) gin.HandlerFunc {
-	return func(context *gin.Context) {
-		isHealthCheck := context.Request.URL.Path == "/healthcheck"
-		secretIsCorrect := context.Query("secret") == secret
-
-		if !isHealthCheck && !secretIsCorrect {
-			_ = context.AbortWithError(http.StatusUnauthorized, fmt.Errorf("Authorization Required"))
-		}
-
-		context.Next()
-	}
-}
-
-func SetupRouter(statusPageClient *statuspage.Client, secret string, componentsStoreChan chan *componentsStore) *gin.Engine {
-	router := gin.Default()
-
-	router.Use(BananaAuthMiddleware(secret))
-	router.Use(InitializeComponentsStore(statusPageClient, componentsStoreChan))
-
-	router.GET("/healthcheck", func(c *gin.Context) {
-		c.Status(http.StatusOK)
-		return
-	})
-
-	router.POST("/", pingdomHandler)
-
-	return router
 }
